Extract color error diagnostic
